[service] move flask app to app.py & Blueprint-ize webhooks
