+++ files/packit.wsgi
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
-
-from packit_service.service.web_hook import app as application
+from packit_service.service.app import application
