+++ packit_service/service/app.py
@@ -1,0 +1,46 @@
+# MIT License
+#
+# Copyright (c) 2018-2019 Red Hat, Inc.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+import logging
+
+from flask import Flask, jsonify
+
+from packit.utils import set_logging
+from packit_service.service.testing_farm_api import blueprint as testing_farm_blueprint
+from packit_service.service.webhooks import blueprint as webhooks_blueprint
+
+
+class PackitReceiver(Flask):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        set_logging(level=logging.DEBUG)
+
+
+application = PackitReceiver(__name__)
+application.register_blueprint(testing_farm_blueprint)
+application.register_blueprint(webhooks_blueprint)
+
+
+@application.route("/healthz", methods=["GET", "HEAD", "POST"])
+def get_health():
+    # TODO: add some interesting stats here
+    return jsonify({"msg": "We are healthy!"})
